src/vdator/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
cinemagoer==2022.2.11
click==8.1.3
discord-markdown==0.4.0
discord.py==2.1.0
emoji==2.2.0
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.1
hunspell==0.5.5
idna==3.4
iso-639==0.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langdetect==1.0.9
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
nltk==3.7
numpy==1.23.5
pathspec==0.10.2
platformdirs==2.5.4
pydash==5.1.2
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tmdbsimple==2.9.1
tomli==2.0.1
tqdm==4.64.1
Unidecode==1.3.6
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
